+++ src/main/java/com/gitlab/cdagaming/craftpresence/utils/discord/rpc/entities/pipe/WindowsPipe.java
@@ -25,12 +25,14 @@
 package com.gitlab.cdagaming.craftpresence.utils.discord.rpc.entities.pipe;
 
 import com.gitlab.cdagaming.craftpresence.ModUtils;
+import com.gitlab.cdagaming.craftpresence.impl.WinRegistry;
 import com.gitlab.cdagaming.craftpresence.utils.discord.rpc.IPCClient;
 import com.gitlab.cdagaming.craftpresence.utils.discord.rpc.entities.Callback;
 import com.gitlab.cdagaming.craftpresence.utils.discord.rpc.entities.Packet;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParseException;
 
+import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.RandomAccessFile;
@@ -54,6 +56,7 @@ public class WindowsPipe extends Pipe {
     }
 
     @Override
+    @SuppressWarnings("BusyWait")
     public Packet read() throws IOException, JsonParseException {
         while ((status == PipeStatus.CONNECTED || status == PipeStatus.CLOSING) && file.length() == 0) {
             try {
@@ -74,17 +77,7 @@ public class WindowsPipe extends Pipe {
 
         file.readFully(d);
 
-        JsonObject packetData = new JsonObject();
-        packetData.addProperty("", new String(d));
-        Packet p = new Packet(op, packetData, ipcClient.getEncoding());
-
-        if (ipcClient.isDebugMode()) {
-            ModUtils.LOG.debugInfo(String.format("Received packet: %s", p.toString()));
-        }
-
-        if (listener != null)
-            listener.onPacketReceived(ipcClient, p);
-        return p;
+        return receive(op, d);
     }
 
     @Override
@@ -97,6 +90,61 @@ public class WindowsPipe extends Pipe {
         send(Packet.OpCode.CLOSE, new JsonObject(), null);
         status = PipeStatus.CLOSED;
         file.close();
+    }
+
+    @Override
+    public void registerApp(String applicationId, String command) {
+        String javaLibraryPath = System.getProperty("java.home");
+        File javaExeFile = new File(javaLibraryPath.split(";")[0] + "/bin/java.exe");
+        File javawExeFile = new File(javaLibraryPath.split(";")[0] + "/bin/javaw.exe");
+        String javaExePath = javaExeFile.exists() ? javaExeFile.getAbsolutePath() : javawExeFile.exists() ? javawExeFile.getAbsolutePath() : null;
+
+        if (javaExePath == null)
+            throw new RuntimeException("Unable to find java path");
+
+        String openCommand;
+
+        if (command != null)
+            openCommand = command;
+        else
+            openCommand = javaExePath;
+
+        String protocolName = "discord-" + applicationId;
+        String protocolDescription = "URL:Run game " + applicationId + " protocol";
+        String keyName = "Software\\Classes\\" + protocolName;
+        String iconKeyName = keyName + "\\DefaultIcon";
+        String commandKeyName = keyName + "\\DefaultIcon";
+
+        try {
+            WinRegistry.createKey(WinRegistry.HKEY_CURRENT_USER, keyName);
+            WinRegistry.writeStringValue(WinRegistry.HKEY_CURRENT_USER, keyName, "", protocolDescription);
+            WinRegistry.writeStringValue(WinRegistry.HKEY_CURRENT_USER, keyName, "URL Protocol", "\0");
+
+            WinRegistry.createKey(WinRegistry.HKEY_CURRENT_USER, iconKeyName);
+            WinRegistry.writeStringValue(WinRegistry.HKEY_CURRENT_USER, iconKeyName, "", javaExePath);
+
+            WinRegistry.createKey(WinRegistry.HKEY_CURRENT_USER, commandKeyName);
+            WinRegistry.writeStringValue(WinRegistry.HKEY_CURRENT_USER, commandKeyName, "", openCommand);
+        } catch (Exception ex) {
+            throw new RuntimeException("Unable to modify Discord registry keys", ex);
+        }
+    }
+
+    @Override
+    public void registerSteamGame(String applicationId, String steamId) {
+        try {
+            String steamPath = WinRegistry.readString(WinRegistry.HKEY_CURRENT_USER, "Software\\\\Valve\\\\Steam", "SteamExe");
+            if (steamPath == null)
+                throw new RuntimeException("Steam exe path not found");
+
+            steamPath = steamPath.replaceAll("/", "\\");
+
+            String command = "\"" + steamPath + "\" steam://rungameid/" + steamId;
+
+            this.registerApp(applicationId, command);
+        } catch (Exception ex) {
+            throw new RuntimeException("Unable to register Steam game", ex);
+        }
     }
 
 }