+++ Changes.md
@@ -12,7 +12,9 @@ _A Detailed Changelog from the last Release is available [here](https://gitlab.c
 *   List Data will no longer update before the config saves (Related to a v1.7.2 change)
 *   `StringUtils#getStringWidth` now properly supports bold characters (Also fixes the release notice position)
 *   All Data in the Color Editor Guis will now synchronize values without needing to push the Enter Key (As such using the Enter Key for syncing has been removed)
-*   Backend Optimizations and Improvements
+*   Added Support for Party Privacy Levels in RPC (See [this commit](https://github.com/discord/discord-rpc/pull/306/))
+*   (Backend-Regression) Added back support for autoRegister into IPC backend
+*   An assortment of backend optimizations and improvements
 
 ### Fixes